Change boolean in pushBlock to flag
